Add files via upload
A priority queue is an extension of a queue data structure which follows the "First in First Out" format meaning the first value in an array is the first to leave and the last value is the last to leave. A priority queue follows the same format however every item is associated with a priority. An item with high priority is dequeued (removed) before an element with low priority even if the element with lower priority was added before. If two items have the same priority, they are served according to their order in the queue